Move msg link embed utility functions to messages.py
